finished the last legacy api

+++ Presentation.Web/Controllers/API/TaskUsageController.cs
@@ -9,29 +9,42 @@ using System.Net.Http.Headers;
 using System.Text;
 using System.Web.Http;
 using Core.ApplicationServices;
+using Core.ApplicationServices.Authorization;
 using Core.DomainModel;
 using Core.DomainModel.ItProject;
 using Core.DomainModel.ItSystemUsage;
 using Core.DomainModel.Organization;
 using Core.DomainServices;
+using Core.DomainServices.Authorization;
 using Newtonsoft.Json.Linq;
 using Presentation.Web.Infrastructure.Attributes;
+using Presentation.Web.Infrastructure.Authorization.Controller.Crud;
 using Presentation.Web.Models;
 using Swashbuckle.Swagger.Annotations;
 
 namespace Presentation.Web.Controllers.API
 {
     [PublicApi]
+    [MigratedToNewAuthorizationContext]
     public class TaskUsageController : GenericHierarchyApiController<TaskUsage, TaskUsageDTO>
     {
         private readonly IGenericRepository<OrganizationUnit> _orgUnitRepository;
         private readonly IGenericRepository<TaskRef> _taskRepository;
 
-        public TaskUsageController(IGenericRepository<TaskUsage> repository, IGenericRepository<OrganizationUnit> orgUnitRepository, IGenericRepository<TaskRef> taskRepository)
-            : base(repository)
+        public TaskUsageController(
+            IGenericRepository<TaskUsage> repository, 
+            IGenericRepository<OrganizationUnit> orgUnitRepository, 
+            IGenericRepository<TaskRef> taskRepository,
+            IAuthorizationContext authorizationContext)
+            : base(repository, authorizationContext)
         {
             _orgUnitRepository = orgUnitRepository;
             _taskRepository = taskRepository;
+        }
+
+        protected override IControllerCrudAuthorization GetCrudAuthorization()
+        {
+            return new ChildEntityCrudAuthorization<TaskUsage>(x => _orgUnitRepository.GetByKey(x.OrgUnitId)?.Organization, base.GetCrudAuthorization());
         }
 
         [HttpGet]
@@ -41,6 +54,11 @@ namespace Presentation.Web.Controllers.API
         {
             try
             {
+                if (GetOrganizationReadAccessLevel(organizationId) < OrganizationDataReadAccessLevel.All)
+                {
+                    return Forbidden();
+                }
+
                 pagingModel.Where(usage => usage.OrgUnitId == orgUnitId);
 
                 if (onlyStarred) pagingModel.Where(usage => usage.Starred);
@@ -75,7 +93,6 @@ namespace Presentation.Web.Controllers.API
                 var orgUnit = _orgUnitRepository.GetByKey(orgUnitId);
                 if (orgUnit == null)
                     return NotFound();
-
                 List<TaskRef> tasks;
                 if (taskId.HasValue)
                 {
@@ -101,14 +118,19 @@ namespace Presentation.Web.Controllers.API
 
                 foreach (var task in tasks)
                 {
-                    Repository.Insert(new TaskUsage()
+                    var taskUsage = new TaskUsage()
                     {
                         OrgUnitId = orgUnitId,
                         TaskRefId = task.Id,
                         ObjectOwner = KitosUser,
                         LastChanged = DateTime.UtcNow,
                         LastChangedByUser = KitosUser
-                    });
+                    };
+                    if (!AllowCreate<TaskUsage>(taskUsage))
+                    {
+                        return Forbidden();
+                    }
+                    Repository.Insert(taskUsage);
                 }
                 Repository.Save();
                 return Ok();
@@ -128,7 +150,10 @@ namespace Presentation.Web.Controllers.API
                 var item = Map<TaskUsageDTO, TaskUsage>(taskUsageDto);
                 item.ObjectOwner = KitosUser;
                 item.LastChangedByUser = KitosUser;
-
+                if (!AllowCreate<TaskUsage>(item))
+                {
+                    return Forbidden();
+                }
                 var savedItem = PostQuery(item);
 
                 return Created(Map(savedItem), new Uri(Request.RequestUri + "/" + savedItem.Id));
@@ -174,6 +199,10 @@ namespace Presentation.Web.Controllers.API
 
                 foreach (var taskUsage in taskUsages)
                 {
+                    if (!AllowDelete(taskUsage))
+                    {
+                        return Forbidden();
+                    }
                     Repository.DeleteByKey(taskUsage.Id);
                 }
                 Repository.Save();
@@ -191,7 +220,10 @@ namespace Presentation.Web.Controllers.API
         {
             try
             {
-                var usages = Repository.Get(usage => usage.OrgUnitId == orgUnitId && usage.Starred == onlyStarred);
+                var usages = Repository
+                    .Get(usage => usage.OrgUnitId == orgUnitId && usage.Starred == onlyStarred)
+                    .Where(AllowRead);
+
                 var dtos = new List<TaskUsageNestedDTO>();
 
                 foreach (var taskUsage in usages)
